requirements.txt
# Research Lens - Python Dependencies

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Database
sqlalchemy>=2.0.0
aiosqlite>=0.19.0

# Authentication
bcrypt>=4.1.0
python-jose[cryptography]>=3.3.0

# LangChain/LangGraph for AI Agents
langgraph>=0.0.40
langchain>=0.1.0
langchain-core>=0.1.0
langchain-google-genai>=1.0.0

# Tavily Search
tavily-python>=0.3.0

# HTTP Client
httpx>=0.26.0

# Environment Variables
python-dotenv>=1.0.0

# Validation
pydantic>=2.5.0
email-validator>=2.1.0

# Langsmith Tracing (optional)
langsmith>=0.0.80
